pyproject.toml
[project]
name = "mkdn"
version = "0.1.0"
description = "CLI Markdown renderer for the terminal. With more features than the rest ;)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "linkify-it-py>=2.1.0",
    "markdown-it-py>=4.0.0",
    "mdit-py-plugins>=0.5.0",
    "pillow>=12.1.1",
    "pygments>=2.19.2",
    "texicode @ git+https://github.com/uukelele/TeXicode.git",
    "typer>=0.24.1",
]

[project.scripts]
mkdn = "mkdn.main:app"

[tool.uv.sources]
texicode = { git = "https://github.com/uukelele/TeXicode.git" }
